gpu: Fix build in 32-bit platforms
Because of the use of u64, this currently fails to build in 32-bit
platforms like wasm. usize is platform dependent but safe enough for
most use cases.

+++ gpu/src/Queue.zig
@@ -71,7 +71,7 @@ pub inline fn writeTexture(
         queue.ptr,
         destination,
         @ptrCast(*const anyopaque, data.ptr),
-        @intCast(u64, data.len) * @sizeOf(std.meta.Elem(@TypeOf(data))),
+        @intCast(usize, data.len) * @sizeOf(std.meta.Elem(@TypeOf(data))),
         data_layout,
         write_size,
     );
